app/requirements.txt
flask==0.12.0
pillow==5.0.0
pg8000==1.11.0
